requirements.txt
# Telegram Bot
python-telegram-bot==21.4
python-dotenv

# Supabase Database - version without strict httpx pin
supabase>=2.0.0,<3.0.0

# Admin Dashboard (FastAPI)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.0
